Goldilocks Monolith
